Wait until the indexer is idle before quitting the IDE
When a refreshnow and shutdownnow file are created together and the refresh triggers Eclipse to update its index, wait until the indexer has been idle for a safe amount of time (currently set to 15 seconds) before quitting.
Also made the logInfo() method of the RefreshNowTask class static so that it can be easily called from other classes.